requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
numpy==2.1.2
onnxruntime==1.19.2
Pillow==11.0.0
